split entities and documents
